Added comments, turned on optimizations
With optimizations set to O3 it crashes, O2 is nice, inlines all
the inlineable functions.  Speed increases from 0.5s to 0.08s.

Starting to make room for padding, and standardizing some naming
conventions. Would really like to make an iterator for the bitmap
class, so many algorithms could be used then.

Signed-off-by: David Post <david@ohyonghao.com>